requirements.txt

viam-sdk==0.43.1

typing-extensions
spidev
mfrc522